+++ lib/home_page/widgets/manga_card.dart
@@ -31,7 +31,7 @@ class MangaCard extends StatelessWidget {
                   color: MuhngaColors.star,
                 );
               },
-              rating: manga['rating'],
+              rating: manga['rating'].toDouble(),
               itemSize: 18),
         ],
       ),
